+++ example/lib/main.dart
@@ -1,126 +1,73 @@
-import 'package:example/test_data.dart';
-import 'package:flutter/material.dart';
+import 'package:fp_state_generator/fp_state_annotation.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+part 'main.fp_state.dart';
+part 'main.freezed.dart';
 
-void main() {
-  runApp(const MyApp());
+@fpState
+class UiState {}
+
+class LState extends UiState {}
+
+class RState extends UiState {}
+
+class R3State extends UiState {}
+
+class R4State extends UiState {}
+
+@freezed
+@fpState
+sealed class FreezedState with _$FreezedState {
+  const FreezedState._();
+  const factory FreezedState.FreezedState(final String user) = FreezedState1;
+  const factory FreezedState.FreezedState2(final String user, final String id) =
+      FreezedState2;
+  const factory FreezedState.FreezedState3(
+      final String user, final bool activate) = FreezedState3;
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+void main(List<String> args) {
+  final state = LState();
+  final match = state.match(
+    lState: (data) => "LState",
+    rState: (data) => "FState",
+    r3State: (R3State data) {},
+    r4State: (R4State data) {},
+  );
+  // match => "LState"
+  final matchOrElse = state.matchOrElse(
+    rState: (data) => "FState",
+    r3State: (R3State data) {},
+    r4State: (R4State data) {},
+    orElse: (data) => "Orlese result",
+  );
+  //  matchOrElse => "Orlese result"
 
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // TRY THIS: Try running your application with "flutter run". You'll see
-        // the application has a purple toolbar. Then, without quitting the app,
-        // try changing the seedColor in the colorScheme below to Colors.green
-        // and then invoke "hot reload" (save your changes or press the "hot
-        // reload" button in a Flutter-supported IDE, or press "r" if you used
-        // the command line to start the app).
-        //
-        // Notice that the counter didn't reset back to zero; the application
-        // state is not lost during the reload. To reset the state, use hot
-        // restart instead.
-        //
-        // This works for code too, not just values: Most code changes can be
-        // tested with just a hot reload.
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
+  final maybeMatch = state.maybeMatch(
+    rState: (data) => "FState",
+    r3State: (R3State data) {},
+    r4State: (R4State data) {},
+  );
 
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
+  // maybeMatch = null
 
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
+  final freezedState = FreezedState.FreezedState2("jack", "1");
+  final freezedMatch = freezedState.match(
+    freezedState1: (data) => "FreezedState",
+    freezedState2: (data) => "FreezedState2 ${data.user},${data.id} ",
+    freezedState3: (data) => "FreezedState3",
+  );
+  // freezedMatch => "FreezedState2 jack,1 "
 
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
+  final freezedMatchOrElse = freezedState.matchOrElse(
+    freezedState1: (data) => "FreezedState",
+    freezedState3: (data) => "FreezedState3",
+    orElse: (data) => "OrElse result",
+  );
+  // freezedMatchOrElse => "OrElse result"
 
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
-    return Scaffold(
-      appBar: AppBar(
-        // TRY THIS: Try changing the color here to a specific color (to
-        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
-        // change color while the other colors stay the same.
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
-      ),
-      body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
-        child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          //
-          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
-          // action in the IDE, or press "p" in the console), to see the
-          // wireframe for each widget.
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
-    );
-  }
+  final freezedMaybeMatch = freezedState.maybeMatch(
+    freezedState1: (data) => "FreezedState",
+    freezedState3: (data) => "FreezedState3",
+  );
+  // freezedMaybeMatch => null
 }