+++ example/lib/main.freezed.dart
@@ -1,0 +1,554 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'main.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+/// @nodoc
+mixin _$FreezedState {
+  String get user => throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(String user) FreezedState,
+    required TResult Function(String user, String id) FreezedState2,
+    required TResult Function(String user, bool activate) FreezedState3,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(String user)? FreezedState,
+    TResult? Function(String user, String id)? FreezedState2,
+    TResult? Function(String user, bool activate)? FreezedState3,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(String user)? FreezedState,
+    TResult Function(String user, String id)? FreezedState2,
+    TResult Function(String user, bool activate)? FreezedState3,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(FreezedState1 value) FreezedState,
+    required TResult Function(FreezedState2 value) FreezedState2,
+    required TResult Function(FreezedState3 value) FreezedState3,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(FreezedState1 value)? FreezedState,
+    TResult? Function(FreezedState2 value)? FreezedState2,
+    TResult? Function(FreezedState3 value)? FreezedState3,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(FreezedState1 value)? FreezedState,
+    TResult Function(FreezedState2 value)? FreezedState2,
+    TResult Function(FreezedState3 value)? FreezedState3,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+
+  @JsonKey(ignore: true)
+  $FreezedStateCopyWith<FreezedState> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $FreezedStateCopyWith<$Res> {
+  factory $FreezedStateCopyWith(
+          FreezedState value, $Res Function(FreezedState) then) =
+      _$FreezedStateCopyWithImpl<$Res, FreezedState>;
+  @useResult
+  $Res call({String user});
+}
+
+/// @nodoc
+class _$FreezedStateCopyWithImpl<$Res, $Val extends FreezedState>
+    implements $FreezedStateCopyWith<$Res> {
+  _$FreezedStateCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? user = null,
+  }) {
+    return _then(_value.copyWith(
+      user: null == user
+          ? _value.user
+          : user // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$FreezedState1ImplCopyWith<$Res>
+    implements $FreezedStateCopyWith<$Res> {
+  factory _$$FreezedState1ImplCopyWith(
+          _$FreezedState1Impl value, $Res Function(_$FreezedState1Impl) then) =
+      __$$FreezedState1ImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String user});
+}
+
+/// @nodoc
+class __$$FreezedState1ImplCopyWithImpl<$Res>
+    extends _$FreezedStateCopyWithImpl<$Res, _$FreezedState1Impl>
+    implements _$$FreezedState1ImplCopyWith<$Res> {
+  __$$FreezedState1ImplCopyWithImpl(
+      _$FreezedState1Impl _value, $Res Function(_$FreezedState1Impl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? user = null,
+  }) {
+    return _then(_$FreezedState1Impl(
+      null == user
+          ? _value.user
+          : user // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$FreezedState1Impl extends FreezedState1 {
+  const _$FreezedState1Impl(this.user) : super._();
+
+  @override
+  final String user;
+
+  @override
+  String toString() {
+    return 'FreezedState.FreezedState(user: $user)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$FreezedState1Impl &&
+            (identical(other.user, user) || other.user == user));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, user);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$FreezedState1ImplCopyWith<_$FreezedState1Impl> get copyWith =>
+      __$$FreezedState1ImplCopyWithImpl<_$FreezedState1Impl>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(String user) FreezedState,
+    required TResult Function(String user, String id) FreezedState2,
+    required TResult Function(String user, bool activate) FreezedState3,
+  }) {
+    return FreezedState(user);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(String user)? FreezedState,
+    TResult? Function(String user, String id)? FreezedState2,
+    TResult? Function(String user, bool activate)? FreezedState3,
+  }) {
+    return FreezedState?.call(user);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(String user)? FreezedState,
+    TResult Function(String user, String id)? FreezedState2,
+    TResult Function(String user, bool activate)? FreezedState3,
+    required TResult orElse(),
+  }) {
+    if (FreezedState != null) {
+      return FreezedState(user);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(FreezedState1 value) FreezedState,
+    required TResult Function(FreezedState2 value) FreezedState2,
+    required TResult Function(FreezedState3 value) FreezedState3,
+  }) {
+    return FreezedState(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(FreezedState1 value)? FreezedState,
+    TResult? Function(FreezedState2 value)? FreezedState2,
+    TResult? Function(FreezedState3 value)? FreezedState3,
+  }) {
+    return FreezedState?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(FreezedState1 value)? FreezedState,
+    TResult Function(FreezedState2 value)? FreezedState2,
+    TResult Function(FreezedState3 value)? FreezedState3,
+    required TResult orElse(),
+  }) {
+    if (FreezedState != null) {
+      return FreezedState(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class FreezedState1 extends FreezedState {
+  const factory FreezedState1(final String user) = _$FreezedState1Impl;
+  const FreezedState1._() : super._();
+
+  @override
+  String get user;
+  @override
+  @JsonKey(ignore: true)
+  _$$FreezedState1ImplCopyWith<_$FreezedState1Impl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class _$$FreezedState2ImplCopyWith<$Res>
+    implements $FreezedStateCopyWith<$Res> {
+  factory _$$FreezedState2ImplCopyWith(
+          _$FreezedState2Impl value, $Res Function(_$FreezedState2Impl) then) =
+      __$$FreezedState2ImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String user, String id});
+}
+
+/// @nodoc
+class __$$FreezedState2ImplCopyWithImpl<$Res>
+    extends _$FreezedStateCopyWithImpl<$Res, _$FreezedState2Impl>
+    implements _$$FreezedState2ImplCopyWith<$Res> {
+  __$$FreezedState2ImplCopyWithImpl(
+      _$FreezedState2Impl _value, $Res Function(_$FreezedState2Impl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? user = null,
+    Object? id = null,
+  }) {
+    return _then(_$FreezedState2Impl(
+      null == user
+          ? _value.user
+          : user // ignore: cast_nullable_to_non_nullable
+              as String,
+      null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$FreezedState2Impl extends FreezedState2 {
+  const _$FreezedState2Impl(this.user, this.id) : super._();
+
+  @override
+  final String user;
+  @override
+  final String id;
+
+  @override
+  String toString() {
+    return 'FreezedState.FreezedState2(user: $user, id: $id)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$FreezedState2Impl &&
+            (identical(other.user, user) || other.user == user) &&
+            (identical(other.id, id) || other.id == id));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, user, id);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$FreezedState2ImplCopyWith<_$FreezedState2Impl> get copyWith =>
+      __$$FreezedState2ImplCopyWithImpl<_$FreezedState2Impl>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(String user) FreezedState,
+    required TResult Function(String user, String id) FreezedState2,
+    required TResult Function(String user, bool activate) FreezedState3,
+  }) {
+    return FreezedState2(user, id);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(String user)? FreezedState,
+    TResult? Function(String user, String id)? FreezedState2,
+    TResult? Function(String user, bool activate)? FreezedState3,
+  }) {
+    return FreezedState2?.call(user, id);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(String user)? FreezedState,
+    TResult Function(String user, String id)? FreezedState2,
+    TResult Function(String user, bool activate)? FreezedState3,
+    required TResult orElse(),
+  }) {
+    if (FreezedState2 != null) {
+      return FreezedState2(user, id);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(FreezedState1 value) FreezedState,
+    required TResult Function(FreezedState2 value) FreezedState2,
+    required TResult Function(FreezedState3 value) FreezedState3,
+  }) {
+    return FreezedState2(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(FreezedState1 value)? FreezedState,
+    TResult? Function(FreezedState2 value)? FreezedState2,
+    TResult? Function(FreezedState3 value)? FreezedState3,
+  }) {
+    return FreezedState2?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(FreezedState1 value)? FreezedState,
+    TResult Function(FreezedState2 value)? FreezedState2,
+    TResult Function(FreezedState3 value)? FreezedState3,
+    required TResult orElse(),
+  }) {
+    if (FreezedState2 != null) {
+      return FreezedState2(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class FreezedState2 extends FreezedState {
+  const factory FreezedState2(final String user, final String id) =
+      _$FreezedState2Impl;
+  const FreezedState2._() : super._();
+
+  @override
+  String get user;
+  String get id;
+  @override
+  @JsonKey(ignore: true)
+  _$$FreezedState2ImplCopyWith<_$FreezedState2Impl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class _$$FreezedState3ImplCopyWith<$Res>
+    implements $FreezedStateCopyWith<$Res> {
+  factory _$$FreezedState3ImplCopyWith(
+          _$FreezedState3Impl value, $Res Function(_$FreezedState3Impl) then) =
+      __$$FreezedState3ImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String user, bool activate});
+}
+
+/// @nodoc
+class __$$FreezedState3ImplCopyWithImpl<$Res>
+    extends _$FreezedStateCopyWithImpl<$Res, _$FreezedState3Impl>
+    implements _$$FreezedState3ImplCopyWith<$Res> {
+  __$$FreezedState3ImplCopyWithImpl(
+      _$FreezedState3Impl _value, $Res Function(_$FreezedState3Impl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? user = null,
+    Object? activate = null,
+  }) {
+    return _then(_$FreezedState3Impl(
+      null == user
+          ? _value.user
+          : user // ignore: cast_nullable_to_non_nullable
+              as String,
+      null == activate
+          ? _value.activate
+          : activate // ignore: cast_nullable_to_non_nullable
+              as bool,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$FreezedState3Impl extends FreezedState3 {
+  const _$FreezedState3Impl(this.user, this.activate) : super._();
+
+  @override
+  final String user;
+  @override
+  final bool activate;
+
+  @override
+  String toString() {
+    return 'FreezedState.FreezedState3(user: $user, activate: $activate)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$FreezedState3Impl &&
+            (identical(other.user, user) || other.user == user) &&
+            (identical(other.activate, activate) ||
+                other.activate == activate));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, user, activate);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$FreezedState3ImplCopyWith<_$FreezedState3Impl> get copyWith =>
+      __$$FreezedState3ImplCopyWithImpl<_$FreezedState3Impl>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(String user) FreezedState,
+    required TResult Function(String user, String id) FreezedState2,
+    required TResult Function(String user, bool activate) FreezedState3,
+  }) {
+    return FreezedState3(user, activate);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(String user)? FreezedState,
+    TResult? Function(String user, String id)? FreezedState2,
+    TResult? Function(String user, bool activate)? FreezedState3,
+  }) {
+    return FreezedState3?.call(user, activate);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(String user)? FreezedState,
+    TResult Function(String user, String id)? FreezedState2,
+    TResult Function(String user, bool activate)? FreezedState3,
+    required TResult orElse(),
+  }) {
+    if (FreezedState3 != null) {
+      return FreezedState3(user, activate);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(FreezedState1 value) FreezedState,
+    required TResult Function(FreezedState2 value) FreezedState2,
+    required TResult Function(FreezedState3 value) FreezedState3,
+  }) {
+    return FreezedState3(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(FreezedState1 value)? FreezedState,
+    TResult? Function(FreezedState2 value)? FreezedState2,
+    TResult? Function(FreezedState3 value)? FreezedState3,
+  }) {
+    return FreezedState3?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(FreezedState1 value)? FreezedState,
+    TResult Function(FreezedState2 value)? FreezedState2,
+    TResult Function(FreezedState3 value)? FreezedState3,
+    required TResult orElse(),
+  }) {
+    if (FreezedState3 != null) {
+      return FreezedState3(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class FreezedState3 extends FreezedState {
+  const factory FreezedState3(final String user, final bool activate) =
+      _$FreezedState3Impl;
+  const FreezedState3._() : super._();
+
+  @override
+  String get user;
+  bool get activate;
+  @override
+  @JsonKey(ignore: true)
+  _$$FreezedState3ImplCopyWith<_$FreezedState3Impl> get copyWith =>
+      throw _privateConstructorUsedError;
+}