+++ lib/fp_state_annotation.dart
@@ -1,3 +1,5 @@
 class FpState {
   const FpState();
 }
+
+const fpState = FpState();