+++ CHANGELOG.md
@@ -1,3 +1,20 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+# 0.0.5
+
+- Support freezed package
+
+# 0.0.4
+
+- Update package info for scores
+
+# 0.0.3
+
+- Update package info
+
+# 0.0.2
+
+- Update package yaml
+
+# 0.0.1
+
+- Initial release!